Fixes #3276. Added flag --name to logs command and moved utils.InferName to a new package
