+++ cmd/deploy/deploy.go
@@ -32,6 +32,7 @@ import (
 	"github.com/okteto/okteto/pkg/cmd/pipeline"
 	"github.com/okteto/okteto/pkg/cmd/stack"
 	"github.com/okteto/okteto/pkg/constants"
+	"github.com/okteto/okteto/pkg/devenvironment"
 	oktetoErrors "github.com/okteto/okteto/pkg/errors"
 	"github.com/okteto/okteto/pkg/format"
 	"github.com/okteto/okteto/pkg/k8s/diverts"
@@ -160,7 +161,7 @@ func Deploy(ctx context.Context) *cobra.Command {
 			}
 			name := options.Name
 			if options.Name == "" {
-				name = utils.InferName(cwd)
+				name = devenvironment.InferName(cwd)
 				if err != nil {
 					return fmt.Errorf("could not infer environment name")
 				}
