+++ cmd/destroy/destroy.go
@@ -30,6 +30,7 @@ import (
 	"github.com/okteto/okteto/cmd/utils/executor"
 	"github.com/okteto/okteto/pkg/config"
 	"github.com/okteto/okteto/pkg/constants"
+	"github.com/okteto/okteto/pkg/devenvironment"
 	oktetoErrors "github.com/okteto/okteto/pkg/errors"
 	"github.com/okteto/okteto/pkg/format"
 
@@ -139,7 +140,7 @@ func Destroy(ctx context.Context) *cobra.Command {
 			}
 			name := options.Name
 			if options.Name == "" {
-				name = utils.InferName(cwd)
+				name = devenvironment.InferName(cwd)
 				if err != nil {
 					return fmt.Errorf("could not infer environment name")
 				}
@@ -218,7 +219,7 @@ func (dc *destroyCommand) runDestroy(ctx context.Context, opts *Options) error {
 		if manifest.Name != "" {
 			opts.Name = manifest.Name
 		} else {
-			opts.Name = utils.InferName(cwd)
+			opts.Name = devenvironment.InferName(cwd)
 		}
 
 	}
